just make it work somehow, todo think of better sync
