+++ src/integratedServer/browserfsShared.ts
@@ -4,14 +4,18 @@ import fs from 'fs'
 import sanitizeFilename from 'sanitize-filename'
 import { oneOf } from '@zardoy/utils'
 import * as browserfs from 'browserfs'
+import { proxy } from 'valtio'
 
 const { GoogleDriveFileSystem } = require('google-drive-browserfs/src/backends/GoogleDrive') // disable type checking
 
-browserfs.install(window)
-const defaultMountablePoints = {
+browserfs.install(globalThis)
+export const defaultMountablePoints = {
   '/world': { fs: 'LocalStorage' }, // will be removed in future
   '/data': { fs: 'IndexedDB' },
   '/resourcepack': { fs: 'InMemory' }, // temporary storage for currently loaded resource pack
+} as Record<string, any>
+if (typeof localStorage === 'undefined') {
+  delete defaultMountablePoints['/world']
 }
 export const configureBrowserFs = (onDone) => {
   browserfs.configure({
@@ -24,26 +28,37 @@ export const configureBrowserFs = (onDone) => {
   })
 }
 
-export const forceCachedDataPaths = {}
-export const forceRedirectPaths = {}
-export const localFsState = {
+export const initialFsState = {
   isReadonly: false,
   syncFs: false,
   inMemorySave: false,
-  remoteBackend: false,
+  inMemorySavePath: '',
   saveLoaded: false,
-  usingIndexFileUrl: null as null | string
+
+  remoteBackend: false,
+  remoteBackendBaseUrl: '',
+  usingIndexFileUrl: '',
+  forceCachedDataPaths: {},
+  forceRedirectPaths: {}
+}
+export const localFsState = {
+  ...initialFsState
 }
 
-globalThis.fs = fs
-//@ts-expect-error
-fs.promises = new Proxy(Object.fromEntries(['readFile', 'writeFile', 'stat', 'mkdir', 'rmdir', 'unlink', 'rename', /* 'copyFile',  */'readdir'].map(key => [key, promisify(fs[key])])), {
+export const currentInternalFsState = proxy({
+  openReadOperations: 0,
+  openWriteOperations: 0,
+  openOperations: 0,
+})
+
+globalThis.fs ??= fs
+const promises = new Proxy(Object.fromEntries(['readFile', 'writeFile', 'stat', 'mkdir', 'rmdir', 'unlink', 'rename', /* 'copyFile',  */'readdir'].map(key => [key, promisify(fs[key])])), {
   get (target, p: string, receiver) {
     if (!target[p]) throw new Error(`Not implemented fs.promises.${p}`)
     return (...args) => {
       // browser fs bug: if path doesn't start with / dirname will return . which would cause infinite loop, so we need to normalize paths
       if (typeof args[0] === 'string' && !args[0].startsWith('/')) args[0] = '/' + args[0]
-      const toRemap = Object.entries(forceRedirectPaths).find(([from]) => args[0].startsWith(from))
+      const toRemap = Object.entries(localFsState.forceRedirectPaths).find(([from]) => args[0].startsWith(from))
       if (toRemap) {
         args[0] = args[0].replace(toRemap[0], toRemap[1])
       }
@@ -51,11 +66,11 @@ fs.promises = new Proxy(Object.fromEntries(['readFile', 'writeFile', 'stat', 'mk
       // todo issue one-time warning (in chat I guess)
       const readonly = localFsState.isReadonly && !(args[0].startsWith('/data') && !localFsState.inMemorySave) // allow copying worlds from external providers such as zip
       if (readonly) {
-        if (oneOf(p, 'readFile', 'writeFile') && forceCachedDataPaths[args[0]]) {
+        if (oneOf(p, 'readFile', 'writeFile') && localFsState.forceCachedDataPaths[args[0]]) {
           if (p === 'readFile') {
-            return Promise.resolve(forceCachedDataPaths[args[0]])
+            return Promise.resolve(localFsState.forceCachedDataPaths[args[0]])
           } else if (p === 'writeFile') {
-            forceCachedDataPaths[args[0]] = args[1]
+            localFsState.forceCachedDataPaths[args[0]] = args[1]
             console.debug('Skipped writing to readonly fs', args[0])
             return Promise.resolve()
           }
@@ -66,22 +81,21 @@ fs.promises = new Proxy(Object.fromEntries(['readFile', 'writeFile', 'stat', 'mk
         args[1] = 'r' // read-only, zipfs throw otherwise
       }
       if (p === 'readFile') {
-        localFsState.openReadOperations++
+        currentInternalFsState.openReadOperations++
       } else if (p === 'writeFile') {
-        localFsState.openWriteOperations++
+        currentInternalFsState.openWriteOperations++
       }
       return target[p](...args).finally(() => {
         if (p === 'readFile') {
-          localFsState.openReadOperations--
+          currentInternalFsState.openReadOperations--
         } else if (p === 'writeFile') {
-          localFsState.openWriteOperations--
+          currentInternalFsState.openWriteOperations--
         }
       })
     }
   }
 })
-//@ts-expect-error
-fs.promises.open = async (...args) => {
+promises.open = async (...args) => {
   //@ts-expect-error
   const fd = await promisify(fs.open)(...args)
   return {
@@ -94,16 +108,16 @@ fs.promises.open = async (...args) => {
         }
 
         if (x === 'read') {
-          localFsState.openReadOperations++
+          currentInternalFsState.openReadOperations++
         } else if (x === 'write' || x === 'close') {
-          localFsState.openWriteOperations++
+          currentInternalFsState.openWriteOperations++
         }
         fs[x](fd, ...args, (err, bytesRead, buffer) => {
           if (x === 'read') {
-            localFsState.openReadOperations--
+            currentInternalFsState.openReadOperations--
           } else if (x === 'write' || x === 'close') {
             // todo that's not correct
-            localFsState.openWriteOperations--
+            currentInternalFsState.openWriteOperations--
           }
           if (err) throw err
           // todo if readonly probably there is no need to open at all (return some mocked version - check reload)?
@@ -130,6 +144,11 @@ fs.promises.open = async (...args) => {
       })
     }
   }
+}
+globalThis.promises = promises
+if (typeof localStorage !== 'undefined') {
+  //@ts-expect-error
+  fs.promises = promises
 }
 
 // for testing purposes, todo move it to core patch
@@ -160,6 +179,28 @@ export const mkdirRecursive = async (path: string) => {
     } catch (err) {
     }
   }
+}
+
+export const mountRemoteFsBackend = async (fsState: typeof localFsState) => {
+  const index = await fetch(fsState.usingIndexFileUrl).then(async (res) => res.json())
+  await new Promise<void>(async resolve => {
+    browserfs.configure({
+      fs: 'MountableFileSystem',
+      options: {
+        ...defaultMountablePoints,
+        '/world': {
+          fs: 'HTTPRequest',
+          options: {
+            index,
+            baseUrl: fsState.remoteBackendBaseUrl
+          }
+        }
+      },
+    }, (e) => {
+      if (e) throw e
+      resolve()
+    })
+  })
 }
 
 export const uniqueFileNameFromWorldName = async (title: string, savePath: string) => {
@@ -276,77 +317,6 @@ export async function removeFileRecursiveAsync (path) {
 }
 
 
-const SUPPORT_WRITE = true
-
-export const openWorldDirectory = async (dragndropHandle?: FileSystemDirectoryHandle) => {
-  //   let _directoryHandle: FileSystemDirectoryHandle
-  //   if (dragndropHandle) {
-  //     _directoryHandle = dragndropHandle
-  //   } else {
-  //     try {
-  //       _directoryHandle = await window.showDirectoryPicker({
-  //         id: 'select-world', // important: this is used to remember user choice (start directory)
-  //       })
-  //     } catch (err) {
-  //       if (err instanceof DOMException && err.name === 'AbortError') return
-  //       throw err
-  //     }
-  //   }
-  //   const directoryHandle = _directoryHandle
-
-  //   const requestResult = SUPPORT_WRITE && !options.preferLoadReadonly ? await directoryHandle.requestPermission?.({ mode: 'readwrite' }) : undefined
-  //   const writeAccess = requestResult === 'granted'
-
-  //   const doContinue = writeAccess || !SUPPORT_WRITE || options.disableLoadPrompts || confirm('Continue in readonly mode?')
-  //   if (!doContinue) return
-  //   await new Promise<void>(resolve => {
-  //     browserfs.configure({
-  //       fs: 'MountableFileSystem',
-  //       options: {
-  //         ...defaultMountablePoints,
-  //         '/world': {
-  //           fs: 'FileSystemAccess',
-  //           options: {
-  //             handle: directoryHandle
-  //           }
-  //         }
-  //       },
-  //     }, (e) => {
-  //       if (e) throw e
-  //       resolve()
-  //     })
-  //   })
-
-  //   localFsState.isReadonly = !writeAccess
-  //   localFsState.syncFs = false
-  //   localFsState.inMemorySave = false
-  //   localFsState.remoteBackend = false
-  //   await loadSave()
-}
-
-const tryToDetectResourcePack = async () => {
-  const askInstall = async () => {
-    // todo investigate browserfs read errors
-    return alert('ATM You can install texturepacks only via options menu.')
-    // if (confirm('Resource pack detected, do you want to install it?')) {
-    //   await installTexturePackFromHandle()
-    // }
-  }
-
-  if (fs.existsSync('/world/pack.mcmeta')) {
-    await askInstall()
-    return true
-  }
-  // const jszip = new JSZip()
-  // let loaded = await jszip.loadAsync(file)
-  // if (loaded.file('pack.mcmeta')) {
-  //   loaded = null
-  //   askInstall()
-  //   return true
-  // }
-  // loaded = null
-}
-
 export const possiblyCleanHandle = (callback = () => { }) => {
   if (!localFsState.saveLoaded) {
     // todo clean handle
@@ -384,48 +354,6 @@ export const collectFilesToCopy = async (basePath: string, safe = false): Promis
   }
   await countFiles('.')
   return result
-}
-
-export const copyFilesAsyncWithProgress = async (pathSrc: string, pathDest: string, throwRootNotExist = true, addMsg = '') => {
-  const stat = await existsViaStats(pathSrc)
-  if (!stat) {
-    if (throwRootNotExist) throw new Error(`Cannot copy. Source directory ${pathSrc} does not exist`)
-    console.debug('source directory does not exist', pathSrc)
-    return
-  }
-  if (!stat.isDirectory()) {
-    await fs.promises.writeFile(pathDest, await fs.promises.readFile(pathSrc) as any)
-    console.debug('copied single file', pathSrc, pathDest)
-    return
-  }
-
-  try {
-    setLoadingScreenStatus('Copying files')
-    let filesCount = 0
-    const countFiles = async (path: string) => {
-      const files = await fs.promises.readdir(path)
-      await Promise.all(files.map(async (file) => {
-        const curPath = join(path, file)
-        const stats = await fs.promises.stat(curPath)
-        if (stats.isDirectory()) {
-          // Recurse
-          await countFiles(curPath)
-        } else {
-          filesCount++
-        }
-      }))
-    }
-    console.debug('Counting files', pathSrc)
-    await countFiles(pathSrc)
-    console.debug('counted', filesCount)
-    let copied = 0
-    await copyFilesAsync(pathSrc, pathDest, (name) => {
-      copied++
-      setLoadingScreenStatus(`Copying files${addMsg} (${copied}/${filesCount}): ${name}`)
-    })
-  } finally {
-    setLoadingScreenStatus(undefined)
-  }
 }
 
 export const existsViaStats = async (path: string) => {
@@ -478,179 +406,4 @@ export const copyFilesAsync = async (pathSrc: string, pathDest: string, fileCopi
       fileCopied?.(curPathDest)
     }
   }))
-}
-
-export const openWorldFromHttpDir = async (fileDescriptorUrls: string[]/*  | undefined */, baseUrlParam) => {
-  // todo try go guess mode
-  let indexFileUrl
-  let index
-  let baseUrl
-  for (const url of fileDescriptorUrls) {
-    let file
-    try {
-      setLoadingScreenStatus(`Trying to get world descriptor from ${new URL(url).host}`)
-      const controller = new AbortController()
-      setTimeout(() => {
-        controller.abort()
-      }, 3000)
-      // eslint-disable-next-line no-await-in-loop
-      const response = await fetch(url, { signal: controller.signal })
-      // eslint-disable-next-line no-await-in-loop
-      file = await response.json()
-    } catch (err) {
-      console.error('Error fetching file descriptor', url, err)
-    }
-    if (!file) continue
-    if (file.baseUrl) {
-      baseUrl = new URL(file.baseUrl, baseUrl).toString()
-      index = file.index
-    } else {
-      index = file
-      baseUrl = baseUrlParam ?? url.split('/').slice(0, -1).join('/')
-    }
-    indexFileUrl = url
-    break
-  }
-  if (!index) throw new Error(`The provided mapDir file is not valid descriptor file! ${fileDescriptorUrls.join(', ')}`)
-  await new Promise<void>(async resolve => {
-    browserfs.configure({
-      fs: 'MountableFileSystem',
-      options: {
-        ...defaultMountablePoints,
-        '/world': {
-          fs: 'HTTPRequest',
-          options: {
-            index,
-            baseUrl
-          }
-        }
-      },
-    }, (e) => {
-      if (e) throw e
-      resolve()
-    })
-  })
-
-  localFsState.saveLoaded = false
-  localFsState.isReadonly = true
-  localFsState.syncFs = false
-  localFsState.inMemorySave = false
-  localFsState.remoteBackend = true
-  localFsState.usingIndexFileUrl = indexFileUrl
-
-  await loadSave()
-}
-
-// todo rename method
-const openWorldZipInner = async (file: File | ArrayBuffer, name = file['name']) => {
-  await new Promise<void>(async resolve => {
-    browserfs.configure({
-      // todo
-      fs: 'MountableFileSystem',
-      options: {
-        ...defaultMountablePoints,
-        '/world': {
-          fs: 'ZipFS',
-          options: {
-            zipData: Buffer.from(file instanceof File ? (await file.arrayBuffer()) : file),
-            name
-          }
-        }
-      },
-    }, (e) => {
-      if (e) throw e
-      resolve()
-    })
-  })
-
-  localFsState.saveLoaded = false
-  localFsState.isReadonly = true
-  localFsState.syncFs = true
-  localFsState.inMemorySave = false
-  localFsState.remoteBackend = false
-
-  if (fs.existsSync('/world/level.dat')) {
-    await loadSave()
-  } else {
-    const dirs = fs.readdirSync('/world')
-    const availableWorlds: string[] = []
-    for (const dir of dirs) {
-      if (fs.existsSync(`/world/${dir}/level.dat`)) {
-        availableWorlds.push(dir)
-      }
-    }
-
-    if (availableWorlds.length === 0) {
-      if (await tryToDetectResourcePack()) return
-      alert('No worlds found in the zip')
-      return
-    }
-
-    if (availableWorlds.length === 1) {
-      await loadSave(`/world/${availableWorlds[0]}`)
-      return
-    }
-
-    alert(`Many (${availableWorlds.length}) worlds found in the zip!`)
-    // todo prompt picker
-    // const selectWorld
-  }
-}
-
-export const openWorldZip = async (...args: Parameters<typeof openWorldZipInner>) => {
-  try {
-    return await openWorldZipInner(...args)
-  } finally {
-    possiblyCleanHandle()
-  }
-}
-
-export const resetLocalStorageWorld = () => {
-  for (const key of Object.keys(localStorage)) {
-    if (/^[\da-fA-F]{8}(?:\b-[\da-fA-F]{4}){3}\b-[\da-fA-F]{12}$/g.test(key) || key === '/') {
-      localStorage.removeItem(key)
-    }
-  }
-}
-
-export const openFilePicker = (specificCase?: 'resourcepack') => {
-  // create and show input picker
-  let picker: HTMLInputElement = document.body.querySelector('input#file-zip-picker')!
-  if (!picker) {
-    picker = document.createElement('input')
-    picker.type = 'file'
-    picker.accept = '.zip'
-
-    picker.addEventListener('change', () => {
-      const file = picker.files?.[0]
-      picker.value = ''
-      if (!file) return
-      if (!file.name.endsWith('.zip')) {
-        const doContinue = confirm(`Are you sure ${file.name.slice(-20)} is .zip file? Only .zip files are supported. Continue?`)
-        if (!doContinue) return
-      }
-      if (specificCase === 'resourcepack') {
-        void installTexturePack(file).catch((err) => {
-          setLoadingScreenStatus(err.message, true)
-        })
-      } else {
-        void openWorldZip(file)
-      }
-    })
-    picker.hidden = true
-    document.body.appendChild(picker)
-  }
-
-  picker.click()
-}
-
-export const resetStateAfterDisconnect = () => {
-  miscUiState.gameLoaded = false
-  miscUiState.loadedDataVersion = null
-  miscUiState.singleplayer = false
-  miscUiState.flyingSquid = false
-  miscUiState.wanOpened = false
-  miscUiState.currentDisplayQr = null
-
-  localFsState.saveLoaded = false
 }
