+++ src/react/SingleplayerProvider.tsx
@@ -3,11 +3,12 @@ import { proxy, subscribe, useSnapshot } from 'valtio'
 import { useEffect, useRef, useState } from 'react'
 import { loadScript } from 'prismarine-viewer/viewer/lib/utils'
 import { fsState, loadSave, longArrayToNumber, readLevelDat } from '../loadSave'
-import { googleDriveGetFileIdFromPath, mountExportFolder, mountGoogleDriveFolder, removeFileRecursiveAsync } from '../browserfs'
+// import { googleDriveGetFileIdFromPath, mountExportFolder, mountGoogleDriveFolder, removeFileRecursiveAsync } from '../browserfs'
 import { hideCurrentModal, showModal } from '../globalState'
 import { haveDirectoryPicker, setLoadingScreenStatus } from '../utils'
 import { exportWorld } from '../builtinCommands'
 import { googleProviderState, useGoogleLogIn, GoogleDriveProvider, isGoogleDriveAvailable, APP_ID } from '../googledrive'
+import { mountExportFolder, removeFileRecursiveAsync } from '../integratedServer/browserfsShared'
 import Singleplayer, { WorldProps } from './Singleplayer'
 import { useIsModalActive } from './utilsApp'
 import { showOptionsModal } from './SelectOption'
