+++ src/shims/fs.js
@@ -1,3 +1,12 @@
 const BrowserFS = require('browserfs')
 
-module.exports = BrowserFS.BFSRequire('fs')
+globalThis.fs ??= BrowserFS.BFSRequire('fs')
+globalThis.fs.promises = new Proxy({}, {
+  get(target, p) {
+    return (...args) => {
+      return globalThis.promises[p](...args)
+    }
+  }
+})
+
+module.exports = globalThis.fs
